pyproject.toml
[tool.poetry]
name = "arcface"
version = "0.1.0"
description = ""
authors = ["desertsniper87 <torshobuet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
tqdm = "^4.66.2"
opencv-python = "^4.9.0.80"
matplotlib = "^3.8.3"
seaborn = "^0.13.2"
torch = "^2.2.1"
torchvision = "^0.17.1"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
